+++ Source/Analytics/Web/src/components/NationalSocietyOverview.js
@@ -4,7 +4,7 @@ import Typography from '@material-ui/core/Typography';
 import { ApplicationInsights } from '@microsoft/applicationinsights-web';
 import CaseReportByHealthRiskTable from "./healthRisk/CaseReportByHealthRiskTable";
 import HealthRiskPerDistrictTable from "./healthRisk/HealthRiskPerDistrictTable";
-import LastWeekTotals from './LastWeekTotals.js';
+import LastWeekTotals from './lastWeekTotals/LastWeekTotals.js';
 import Map from "./Map.js";
 import CBSNavigation from './Navigation/CBSNavigation';
